feat: pages cleanup
